removed old not used code for handle remote tree ridbags
