+++ client/src/main/java/com/orientechnologies/orient/client/remote/message/OSBTFetchEntriesMajorResponse.java
@@ -28,8 +28,8 @@ import java.util.Map.Entry;
 import com.orientechnologies.common.serialization.types.OBinarySerializer;
 import com.orientechnologies.common.serialization.types.OIntegerSerializer;
 import com.orientechnologies.orient.client.remote.OBinaryResponse;
-import com.orientechnologies.orient.client.remote.OSBTreeBonsaiRemote;
 import com.orientechnologies.orient.client.remote.OStorageRemoteSession;
+import com.orientechnologies.orient.client.remote.TreeEntry;
 import com.orientechnologies.orient.core.serialization.serializer.record.ORecordSerializer;
 import com.orientechnologies.orient.enterprise.channel.binary.OChannelDataInput;
 import com.orientechnologies.orient.enterprise.channel.binary.OChannelDataOutput;
@@ -63,7 +63,7 @@ public class OSBTFetchEntriesMajorResponse<K, V> implements OBinaryResponse {
       offset += keySerializer.getObjectSize(stream, offset);
       final V resultValue = valueSerializer.deserialize(stream, offset);
       offset += valueSerializer.getObjectSize(stream, offset);
-      list.add(new OSBTreeBonsaiRemote.TreeEntry<K, V>(resultKey, resultValue));
+      list.add(new TreeEntry<K, V>(resultKey, resultValue));
     }
   }
